Refactor: store폴더 만들어서 redux를 따로 분리
